requirements.txt
autopep8==1.4.4
cffi==1.13.2
coverage==4.5.4
cryptography==2.8
Django==2.2.7
django-templated-mail==1.1.1
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.9
jwt==0.6.1
mccabe==0.6.1
psycopg2==2.7.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
